写了个噪声生成器

+++ src/main/java/yee/pltision/game/world/entity/DynamicPosition.java
@@ -1,8 +1,8 @@
 package yee.pltision.game.world.entity;
 
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 
 public interface DynamicPosition {
-    Vector3f getPosition();
+    Vector3d getPosition();
 
 }
